fix(outline): keep the client editor out of the server-only specializer bundle
The outline editor (a client component) imports resetConfigForType from the
specializer, but the specializer statically imports sanitizeProceduralSkillOutline
from outline-generator, which transitively pulls in the fs-backed prompt loader.
That dragged `fs` into the client bundle and broke `next build`
(Module not found: Can't resolve 'fs').

Move the pure, client-safe helpers (computeIntentHash, hasRequiredConfig,
resetConfigForType) into outline-specializer-helpers.ts with no server imports.
The specializer imports/re-exports them; the editor imports from the helpers
module directly. Production build passes; 25 unit tests unchanged.
